+++ test/conversation_test.exs
@@ -32,16 +32,22 @@ defmodule ConversationTest do
     %{conn: conn, map_conn: map_conn}
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: list conversation models", %{conn: conn, map_conn: map_conn} do
-    assert {:ok, models} = ExTypesense.list_models()
-    assert length(models) >= 0
-    assert {:ok, _} = ExTypesense.list_models([])
-    assert {:ok, _} = ExTypesense.list_models(conn: conn)
-    assert {:ok, _} = ExTypesense.list_models(conn: map_conn)
+    case ExTypesense.list_models() do
+      {:ok, []} ->
+        assert {:ok, []} = ExTypesense.list_models([])
+        assert {:ok, []} = ExTypesense.list_models(conn: conn)
+        assert {:ok, []} = ExTypesense.list_models(conn: map_conn)
+
+      {:ok, [first | _]} when is_struct(first, NLSearchModelSchema) ->
+        assert {:ok, _} = ExTypesense.list_models([])
+        assert {:ok, _} = ExTypesense.list_models(conn: conn)
+        assert {:ok, _} = ExTypesense.list_models(conn: map_conn)
+    end
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "error: get a non-existent conversation model", %{conn: conn, map_conn: map_conn} do
     assert {:error, %ApiResponse{message: "Model not found"}} =
              ExTypesense.get_model("non-existent")
@@ -51,7 +57,7 @@ defmodule ConversationTest do
     assert {:error, _} = ExTypesense.get_model("xyz", conn: map_conn)
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: delete a conversation model", %{conn: conn, map_conn: map_conn} do
     assert {:error, %ApiResponse{message: "Model not found"}} =
              ExTypesense.delete_model("non-existent")
@@ -61,7 +67,7 @@ defmodule ConversationTest do
     assert {:error, _} = ExTypesense.delete_model("xyz", conn: map_conn)
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "error: create a conversation model with incorrect API key", %{
     conn: conn,
     map_conn: map_conn
@@ -92,7 +98,7 @@ defmodule ConversationTest do
     assert {:error, _} = ExTypesense.create_model(body, conn: map_conn)
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "error: update a conversation model with incorrect API key", %{
     conn: conn,
     map_conn: map_conn