+++ test/debug_test.exs
@@ -2,7 +2,6 @@ defmodule DebugTest do
   use ExUnit.Case, async: true
 
   alias OpenApiTypesense.Connection
-  alias OpenApiTypesense.Debug
 
   setup_all do
     conn = Connection.new()
@@ -11,9 +10,9 @@ defmodule DebugTest do
     %{conn: conn, map_conn: map_conn}
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
-  test "debug", %{conn: conn, map_conn: map_conn} do
-    assert {:ok, %Debug{version: _}} = ExTypesense.debug()
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  test "debug (v29.0)", %{conn: conn, map_conn: map_conn} do
+    assert {:ok, %{"state" => 1, "version" => _version}} = ExTypesense.debug()
     assert {:ok, _} = ExTypesense.debug([])
     assert {:ok, _} = ExTypesense.debug(conn: conn)
     assert {:ok, _} = ExTypesense.debug(conn: map_conn)