+++ test/stopwords_test.exs
@@ -17,14 +17,38 @@ defmodule StopwordsTest do
       stopwords
       |> Enum.each(fn stopword ->
         stop_id = stopword.id
-        {:ok, %OpenApiTypesense.Stopwords{id: ^stop_id}} = ExTypesense.delete_stopword(stop_id)
+        {:ok, _} = ExTypesense.delete_stopword(stop_id)
       end)
     end)
 
     %{conn: conn, map_conn: map_conn}
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  @tag ["29.0": true]
+  test "success (v29.0): get stopword", %{conn: conn, map_conn: map_conn} do
+    stop_id = "stopword_set_countries"
+
+    body =
+      %{
+        "stopwords" => ["Germany", "France", "Italy", "United States"],
+        "locale" => "en"
+      }
+
+    assert {:ok, %StopwordsSetSchema{id: ^stop_id}} = ExTypesense.upsert_stopword(stop_id, body)
+    assert {:ok, _} = ExTypesense.upsert_stopword(stop_id, body, [])
+    assert {:ok, _} = ExTypesense.upsert_stopword(stop_id, body, conn: conn)
+    assert {:ok, _} = ExTypesense.upsert_stopword(stop_id, body, conn: map_conn)
+
+    assert {:ok, %StopwordsSetRetrieveSchema{stopwords: %{"id" => ^stop_id}}} =
+             ExTypesense.get_stopword(stop_id)
+
+    assert {:ok, _} = ExTypesense.get_stopword(stop_id)
+    assert {:ok, _} = ExTypesense.get_stopword(stop_id, [])
+    assert {:ok, _} = ExTypesense.get_stopword(stop_id, conn: conn)
+    assert {:ok, _} = ExTypesense.get_stopword(stop_id, conn: map_conn)
+  end
+
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: get stopword", %{conn: conn, map_conn: map_conn} do
     stop_id = "stopword_set_countries"
 
@@ -39,8 +63,13 @@ defmodule StopwordsTest do
     assert {:ok, _} = ExTypesense.upsert_stopword(stop_id, body, conn: conn)
     assert {:ok, _} = ExTypesense.upsert_stopword(stop_id, body, conn: map_conn)
 
-    assert {:ok, %StopwordsSetRetrieveSchema{stopwords: %{id: ^stop_id}}} =
-             ExTypesense.get_stopword(stop_id)
+    case ExTypesense.get_stopword(stop_id) do
+      {:ok, %StopwordsSetRetrieveSchema{stopwords: %{id: id}}} ->
+        assert stop_id === id
+
+      {:ok, %StopwordsSetRetrieveSchema{stopwords: %{"id" => id}}} ->
+        assert stop_id === id
+    end
 
     assert {:ok, _} = ExTypesense.get_stopword(stop_id)
     assert {:ok, _} = ExTypesense.get_stopword(stop_id, [])
@@ -48,7 +77,7 @@ defmodule StopwordsTest do
     assert {:ok, _} = ExTypesense.get_stopword(stop_id, conn: map_conn)
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: list all stopwords", %{conn: conn, map_conn: map_conn} do
     assert {:ok, %StopwordsSetsRetrieveAllSchema{}} = ExTypesense.list_stopwords()
     assert {:ok, _} = ExTypesense.list_stopwords([])
@@ -56,7 +85,7 @@ defmodule StopwordsTest do
     assert {:ok, _} = ExTypesense.list_stopwords(conn: map_conn)
   end
 
-  @tag ["28.0": true, "27.1": true, "27.0": true, "26.0": true]
+  @tag ["29.0": true, "28.0": true, "27.1": true, "27.0": true, "26.0": true]
   test "success: delete stopword", %{conn: conn, map_conn: map_conn} do
     stop_id = "stopword_set_countries"
 
@@ -67,7 +96,14 @@ defmodule StopwordsTest do
       }
 
     assert {:ok, %StopwordsSetSchema{id: ^stop_id}} = ExTypesense.upsert_stopword(stop_id, body)
-    assert {:ok, %OpenApiTypesense.Stopwords{id: ^stop_id}} = ExTypesense.delete_stopword(stop_id)
+
+    case ExTypesense.delete_stopword(stop_id) do
+      {:ok, map} when is_map(map) ->
+        assert {:ok, map} === {:ok, %{"id" => stop_id}}
+
+      {:ok, struct} when is_struct(struct) ->
+        assert {:ok, struct} === {:ok, %OpenApiTypesense.Stopwords{id: stop_id}}
+    end
 
     assert {:error, %ApiResponse{message: "Stopword `stopword_set_countries` not found."}} =
              ExTypesense.delete_stopword(stop_id, [])