+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## major.minor.patch (yyyy.mm.dd)
 
+## 2.1.0 (2026.04.13)
+
+### Chore
+
+* Update compatibility to include Typesense v29.0
+* Added `ExTypesense.NaturalLanguage` operation
+* Added `ExTypesense.flush_analytics/1`
+* Added `ExTypesense.get_analytics_status/1`
+* Added `ExTypesense.get_analytics_events/1`
+* Added tests for new operations
+
+### Removed
+
+* `Operations.config/2`. Use `Operations.toggle_slow_request_log/2`
+
 ## 2.0.0 (2025.06.01)
 
 ### Changed